Update shared/drivers to mp2

+++ shared/drivers/register_access.hh
@@ -4,7 +4,7 @@
 
 namespace mdrivlib
 {
-using regsize_t = unsigned long; // 32bits on Cortex-M, 64-bits on x86_64 and ARM64
+using regsize_t = uint32_t; // always 32bits
 
 // Mask Helpers
 // from github.com/kensmith/cortex-from-scratch
@@ -27,8 +27,7 @@ static_assert(mask_v<2, 4> == 0b111100);
 // first_bit():
 // Returns the position of the first set bit (LSB = 0)
 // Returns greater than the register size if no bits are set
-constexpr regsize_t first_bit(regsize_t mask)
-{
+constexpr regsize_t first_bit(regsize_t mask) {
 	regsize_t fb = 0;
 	if (mask == 0)
 		return sizeof(regsize_t) * 8;
@@ -49,20 +48,16 @@ struct RegisterBits {
 	static constexpr regsize_t Mask = mask;
 	static constexpr regsize_t offset = first_bit(mask);
 
-	static regsize_t read()
-	{
+	static regsize_t read() {
 		return AccessPolicyT::read(reinterpret_cast<volatile regsize_t *>(address), offset, mask);
 	}
-	static void write(regsize_t val)
-	{
+	static void write(regsize_t val) {
 		AccessPolicyT::write(reinterpret_cast<volatile regsize_t *>(address), offset, mask, val);
 	}
-	static void set()
-	{
+	static void set() {
 		AccessPolicyT::set(reinterpret_cast<volatile regsize_t *>(address), mask);
 	}
-	static void clear()
-	{
+	static void clear() {
 		AccessPolicyT::clear(reinterpret_cast<volatile regsize_t *>(address), mask);
 	}
 };
@@ -74,46 +69,42 @@ using RegisterSection = RegisterBits<AccessPolicyT, address, mask_v<offset, widt
 
 // Null register (periph does not exist)
 struct NonexistantRegister {
-	static regsize_t read()
-	{
+	static regsize_t read() {
 		return 0;
 	}
-	static void write(regsize_t val) {}
-	static void set() {}
-	static void clear() {}
+	static void write(regsize_t val) {
+	}
+	static void set() {
+	}
+	static void clear() {
+	}
 };
 
 // Access Policies:
 
 struct ReadOnly {
-	static regsize_t read(volatile regsize_t *address, regsize_t offset, regsize_t mask)
-	{
+	static regsize_t read(volatile regsize_t *address, regsize_t offset, regsize_t mask) {
 		return (*address & mask) >> offset;
 	}
 };
 
 struct WriteOnly {
-	static void write(volatile regsize_t *address, regsize_t offset, regsize_t mask, regsize_t val)
-	{
+	static void write(volatile regsize_t *address, regsize_t offset, regsize_t mask, regsize_t val) {
 		*address = ((val << offset) & mask);
 	}
-	static void set(volatile regsize_t *address, regsize_t mask)
-	{
+	static void set(volatile regsize_t *address, regsize_t mask) {
 		*address = mask;
 	}
 };
 
 struct ReadWrite : ReadOnly {
-	static void write(volatile regsize_t *address, regsize_t offset, regsize_t mask, regsize_t val)
-	{
+	static void write(volatile regsize_t *address, regsize_t offset, regsize_t mask, regsize_t val) {
 		*address = (*address & ~mask) | ((val << offset) & mask);
 	}
-	static void set(volatile regsize_t *address, regsize_t mask)
-	{
+	static void set(volatile regsize_t *address, regsize_t mask) {
 		*address = *address | mask;
 	}
-	static void clear(volatile regsize_t *address, regsize_t mask)
-	{
+	static void clear(volatile regsize_t *address, regsize_t mask) {
 		*address = *address & ~mask;
 	}
 };
@@ -122,12 +113,10 @@ struct ReadWrite : ReadOnly {
 // Access to Set/Clear is always WriteOnly.
 template<regsize_t SetAddress, regsize_t SetMask, regsize_t ClearAddress, regsize_t ClearMask>
 struct RegisterDualSetClear {
-	static void set()
-	{
+	static void set() {
 		WriteOnly::set(reinterpret_cast<volatile regsize_t *>(SetAddress), SetMask);
 	}
-	static void clear()
-	{
+	static void clear() {
 		WriteOnly::set(reinterpret_cast<volatile regsize_t *>(ClearAddress), ClearMask);
 	}
 };
